Updated retriever.py and tests.json
Fixed erroneus json file and updated retriever.py to process a query
using a different method.

+++ retriever.py
@@ -18,7 +18,7 @@ class Test:
         solutions: the url of the test's solutions
         """
         self.rawdata = data
-        self.year = data[0]
+        self.year = int(data[0])
         self.event = data[1].replace(":","-")
         self.type = data[2]
         self.test = data[3]
@@ -28,7 +28,7 @@ class Test:
 class Subject:
     def __init__(self, data):
         self.rawdata = data
-        self.name = data["aliases"][0]
+        self.name = data['aliases'][0]
         self.aliases = data["aliases"]#[1:len(data["aliases"])]
         self.tests = [Test(a) for a in data["tests"]]
         self.testdict = dict()
@@ -36,16 +36,17 @@ class Subject:
             pass
         self.cache_tests()
     def cache_tests(self):
-        print("Caching Tests...")
+        print('Caching Tests (' + self.name + ')...')
         if not os.path.exists(self.name+"/"):
             os.makedirs(self.name + "/")
         for t in self.tests:
-            if not os.path.isfile(self.name + "/" + t.testname):
-                testdata = urllib.request.urlopen(t.test).read() if t.test != "" else b''
+            print('Attempting: ' + str(t.year) + ' ' + t.event)
+            if (not os.path.isfile(self.name + "/" + t.testname)) and t.test != '':
+                testdata = urllib.request.urlopen(t.test).read()
                 testfile = open(self.name + "/" + t.testname, "wb")
                 testfile.write(testdata)
                 testfile.close()
-            if not os.path.isfile(self.name + "/" + t.solutionsname):
+            if (not os.path.isfile(self.name + "/" + t.solutionsname)) and t.solutions != '':
                 testdata = urllib.request.urlopen(t.solutions).read() if t.solutions != "" else b''
                 testfile = open(self.name + "/" + t.solutionsname, "wb")
                 testfile.write(testdata)
@@ -70,90 +71,46 @@ def indexOf(s1,arr):
         if s1 in s:
             return arr.index(s)
     return None
-testdata = json.loads(open("tests.json", "r").read())
 
-calculusdata = testdata["calculus"]
-precalculusdata = testdata["precalculus"]
-algebradata = testdata["algebra"]
-geometrydata = testdata["geometry"]
+def replace_subject(string, subject):
+    for a in subject.aliases:
+        if (' ' + a) in string:
+            print('Found ' + a + ' in query, replacing...')
+            return string.replace(a, subject.name)
+    return string
 
-calculus = Subject(calculusdata)
-precalc = Subject(precalculusdata)
-algebra = Subject(algebradata)
-geometry = Subject(geometrydata)
-
-subjects = [calculus,precalc,algebra,geometry]
-
-print(calculus.as_string())
-print(precalc.as_string())
-print(algebra.as_string())
-print(geometry.as_string())
-
-class InterestObject:
-    def __init__(self, desc = '', index = 0, data = None):
-        self.description = desc
-        self.index = index
-        self.data = data
-
-def run():
-    query = input("Request: ")
-    sentence = query.split(" ")
-    similarwords = {"test":["tests","test","quiz","quizzes"]}
-    keywords = ['all','and','from','through', 'to', 'with', 'without']
-    interest = []
-    for word in sentence:
+def run_main():
+    with open('tests.json', 'r') as loadfile:
+        testdata = json.load(loadfile)
+        subjects = [Subject(_s_) for _s_ in testdata.values()]
+        to_download = []
+        query = ' ' + input('query: ').strip()
         for s in subjects:
-            for a in s.aliases:
-                if s.name == 'calculus' or s.name == 'precalculus':
-                    print("Discerning between precalc and calc")
-                    print("Chance of " + s.name + ': ' + str(str_ratio(word, a)))
-                    if str_ratio(word, a) >= 0.90:
-                        print('\nreplacing\n')
-                        query = query.replace(word, s.name)
+            query = replace_subject(query, s)
+        query = query.replace(' tests ', ' ').replace(' test ', ' ').replace(' from ', ' ').replace(' with ', ' ')
+        query = query.strip()
+        if 'and' in query:
+            query = query.split(' and ')
+        else:
+            query = [query]
+        print(str(query))
+        for subquery in query:
+            subquery = ' ' + subquery + ' '
+            for s in subjects:
+                if ' ' + s.name + ' ' in subquery:
+                    numbers = [int(i) for i in subquery.split() if i.isdigit()]
+                    if len(numbers) == 0:
+                        print(s.name + ' no numbers, continuing')
                         continue
-                if str_ratio(word, a) >= 0.7:
-                    query = query.replace(word, s.name)
-    for s in subjects:
-        for a in similarwords["test"]:
-            if s.name + ' ' + a in query:
-                query = query.replace(s.name + ' ' + a, s.name)
-    sentence = query.split(' ')
-    print("First process: " + query)
-    similaritydict = {s.name: max([str_ratio(w, a) for a in s.aliases for w in sentence]) for s in subjects}
-    for word in sentence:
-        length = len(sentence)
-        index = sentence.index(word)
-        if word.lower() in keywords:
-            if word.lower() == "all":
-                interest.append(InterestObject('all', index, sentence[index:]))
-                continue
-            if word.lower() == "and" and not index == length - 1:
-                interest.append(InterestObject('and', index, sentence[:index]+sentence[index+1:]))
-                continue
-            if (index < (length - 1) and not index == 0):
-                interest.append(InterestObject(word.lower(), index, (sentence[index-1],sentence[index+1])))
-
-    similaritydict = {s.name: max([str_ratio(w, a) for a in s.aliases for w in sentence]) for s in subjects}
-
-    print("Similarity:")
-    [print("\t" + a + ': ' + str(b)) for a,b in similaritydict.items()]
-
-    for a,b in similaritydict.items():
-        if(b >= 0.7):
-            interest.append(InterestObject(a + " found in sentence", indexOf(a, sentence)))
-
-    print("Interests")
-    [print("\t" + str(i.index) + ": " + i.description + '    -    ' + str(i.data)) for i in interest]
-
-    for n in interest:
-        if 'comma' in n.description.lower():
-            pass
-        if 'all' in n.description.lower():
-            pass
-        if 'and' in n.description.lower():
-            pass
-
+                    elif len(numbers) == 1:
+                        #only one year
+                        print(s.name + ' one year requested')
+                    elif len(numbers) == 2:
+                        #range of years
+                        print(s.name + ' range of years requested')
+                    else:
+                        print(s.name + ' more than two numbers given, passing')
+                        pass
 if __name__ == '__main__':
-    run()
-
+    run_main()
